25/12/27 신뢰게이지 계산 메소드 구현 (디테일 구현중)

+++ src/main/java/com/dealOn/common/model/vo/ReviewVO.java
@@ -17,7 +17,7 @@ public class ReviewVO {
 	private int buyerNo;
 	private String reviewText;
 	private String status;
-	private float rateScore;
+	private int rateScore;
 	private Date createDate;
 	private String sellerNickname;
 	private String buyerNickname;
